apps/LibriSpeech/requirements.txt
flask
hydra-core
jupyter
librosa
matplotlib
numpy
optuna
pytest
python-levenshtein
pytorch-lightning==1.7.0
scipy
scikit-learn
sox
torch
torchaudio
torchelastic
tqdm
wget
git+https://github.com/romesco/hydra-lightning/#subdirectory=hydra-configs-pytorch-lightning
pymongo
hydra-core
librosa
pandas
torchmetrics==0.7.3